fix(sanity): remove duplicate customer schema
